requirements.txt
# A股股票均线监控依赖

# 核心依赖
akshare>=1.12.0        # A股数据获取（数据源：新浪/东方财富）
pandas>=2.0.0          # 数据处理
pyyaml>=6.0            # 配置文件解析

# 通知
requests>=2.28.0       # HTTP 请求

# 定时任务
schedule>=1.2.0        # 定时任务调度

# 工具
python-dotenv>=1.0.0   # 环境变量加载
